requirements.txt
numpy>=1.26
opencv-python>=4.10
Pillow>=10.0
PyYAML>=6.0
rapidocr>=3.8
onnxruntime>=1.24
requests>=2.32